AOC Day 1 in R without tests
I decided to do the AOC project pretty late (10:30pm on day 1).
Although I'd like to use this opportunity to learn a new language
(and I may yet return and do this over in something like Elixir),
I'd rather get this done and stick with it this year. So I did a
quick and dirty answer on day 1 in my language of choice, R, without
tests. I definitely will com back and add tests.

Some things of note:
 - recursion is a bit weird in R and I like solving this problem using
   vectors. The thing is, I'm tired, so I used a `while` loop possibly
   for the first time ever in R. I don't love this solution and want
   to think about it more over the coming days. It feels non-idiomatic.
 - on the other hand, I love using `purrr::keep` to simplify the logic
   around not dealing with negative fuels and handling a "stop"
   conditions in a complex way.
 - I was surprised this problem ran fast using R with this particular
   code. If I do come back and do this without something like a while
   loop (seems possible?), I'd be curious if it performs better than
   this version.
 - I was lazy and did not make this a #! /bin/R script but that's
   probably how I'd like to do the rest of these problems.
 - as noted in code comments, using `httr::GET` or `url` to fetch data
   would have required dealing with some oauth stuff and that didn't
   seem worth it so I just copied and pasted the input data.