Day 3 was really hard
So this was a challenge for me. At first, I thought I was clever and
came up with a way of getting each of the coordinates for a move and
thought of my inner join strategy. That's when I realized it was not
just the "hops" but _each step_ I had to care about. That caused me to
rewrite my whole approach. I reached back for a while loop again because
recursion in R feels hard and wrong to me. I'm annoyed I couldn't think
of a way to use a different iterator effectively, but the idea of
"consuming" an iterable and having prior results in R is just unnatural
compared to other languages.

What was great though was doing the work for part 1 the "hard way"
actually made part 2 super easy. So whereas part 1 took 90 mins, part 2
took less than 10, mostly from lazy errors because I'm tired.

Nothing too interesting in the code itself from today.