Day 4 regex is cool
So I went down a really slow, bad path here with pmap etc for the
first part, and then as I thought about how to solve the problem more
carefully for part 2, I realized that regex
was, in fact, the answer. I found the instrutions for part 2 confusing,
but ultimately after talking outloud to coworkers realized what was needed.

Most notable for me was the use of the capture group `\\1` directly in the
matching pattern for my regex which I'm not sure I knew you could do.