Oh … oh my
Finished this one under the gun.

I found this super hard, not the least of which was because keeping track of my position in R when the data is 0-indexed and R is 1-indexed was really crumby. I would say at least an hour was spend debugging bugs related ot that.

I also needed help from a friend to understand what the jump codes really meant— I didn’t understand the pointer itself was moving rather than modifying the value of the current pointer and then re-running the instruction set from that position in the data.